move observables, add submodule docs
