+++ pennylane/ops/qubit/__init__.py
@@ -22,7 +22,6 @@ from .qchem_ops import *
 from .arbitrary_ops import *
 from .arthimetic_ops import *
 from .state_preparation import *
-from .observables import *
 from .other import *
